🔧 [#113] Address Devin's PR review comments
- Fix FQCN violations in routes/api.php (use imports, not inline)
- Change .env.example CLOCK_SIMULATION_ENABLED default to false
- Fix docstring in OperatingUnit.php (nowInBusinessTz)
- Fix docstring in EmploymentPeriod.php (nowInBusinessTz)
- Extract ClockBadge logic to use-clock-badge.ts custom hook
- Fix ScheduleResource time consistency with reference_date
