+++ code/api/.env.example
@@ -94,4 +94,4 @@ DEV_LOGIN_ALLOWED_ENVIRONMENTS=dev,devtest
 # When true, /api/v1/devtools/clock endpoints become available.
 # The ClockBadge in the header shows when clock is in "simulated" mode.
 # CRITICAL: Must be false in production.
-CLOCK_SIMULATION_ENABLED=true
+CLOCK_SIMULATION_ENABLED=false
