+++ code/api/app/Models/OperatingUnit.php
@@ -117,7 +117,7 @@ class OperatingUnit extends Model
     /**
      * Scope to filter active events (within date range).
      *
-     * @param  DateTimeInterface  $referenceDate  The date to compare against (use ApplicationClock::todayInBusinessTz())
+     * @param  DateTimeInterface  $referenceDate  The date to compare against (use ApplicationClock::nowInBusinessTz())
      */
     public function scopeActiveEvents($query, DateTimeInterface $referenceDate)
     {
